+++ DemolisherWeapon/Render/FinalRender.h
@@ -103,7 +103,7 @@ private:
 
 	CPrimitive m_drawSpace;
 	CPrimitive::SVertex m_vertex[4];
-	int m_index[4] = { 0,1,2,3 };
+	unsigned long m_index[4] = { 0,1,2,3 };
 
 	//•`‰æ”ÍˆÍ
 	CVector2 m_screen_min;
